rebased and move sql related code to mysql package
