+++ tools/monitoring/mysql/database.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package log_parser
+package mysql
 
 import (
 	"database/sql"
